moved github and bitbucket input packages under scm package.
