+++ src/main/java/io/jenkins/plugins/synopsys/security/scan/service/ScannerArgumentService.java
@@ -13,10 +13,10 @@ import io.jenkins.plugins.synopsys.security.scan.global.Utility;
 import io.jenkins.plugins.synopsys.security.scan.global.enums.SecurityProduct;
 import io.jenkins.plugins.synopsys.security.scan.input.BridgeInput;
 import io.jenkins.plugins.synopsys.security.scan.input.NetworkAirGap;
-import io.jenkins.plugins.synopsys.security.scan.input.bitbucket.Bitbucket;
+import io.jenkins.plugins.synopsys.security.scan.input.scm.bitbucket.Bitbucket;
 import io.jenkins.plugins.synopsys.security.scan.input.blackduck.BlackDuck;
 import io.jenkins.plugins.synopsys.security.scan.input.coverity.Coverity;
-import io.jenkins.plugins.synopsys.security.scan.input.github.Github;
+import io.jenkins.plugins.synopsys.security.scan.input.scm.github.Github;
 import io.jenkins.plugins.synopsys.security.scan.input.polaris.Polaris;
 import io.jenkins.plugins.synopsys.security.scan.input.scm.gitlab.Gitlab;
 import io.jenkins.plugins.synopsys.security.scan.service.scan.ScanParametersService;
